requirements.txt
appdirs==1.4.3
diff-match-patch==20121119
Django==1.10.2
django-bootstrap-toolkit==2.15.0
django-bootstrap3==7.1.0
django-geojson==2.10.0
django-geoposition==0.3.0
django-google-maps==0.6.0
django-import-export==0.5.1
django-jsonfield==1.0.1
django-leaflet==0.19.0
djangorestframework==3.5.3
gunicorn==19.7.1
jsonfield==2.0.1
packaging==16.8
psycopg2==2.6.1
pyparsing==2.2.0
six==1.10.0
tablib==0.11.4
whitenoise==3.3.0
